% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group2.R
\name{Bootstrap_function}
\alias{Bootstrap_function}
\title{Bootstrap Confidence Interval for Logistic Regression Coefficients}
\usage{
Bootstrap_function(alpha, B = 20, X, y)
}
\arguments{
\item{alpha}{The significance level used to calculate the confidence interval.
Typically set to 0.05 for a 95% confidence interval.}

\item{B}{The number of bootstrap samples to generate. Default is 20.}

\item{X}{A \code{vector} or \code{matrix} representing the predictor variables.}

\item{y}{A \code{vector} representing the binary response variable.}
}
\value{
Returns a \code{vector} containing the lower and upper bounds of the
confidence interval for the mean of the bootstrapped coefficient estimates.
}
\description{
Computes bootstrap confidence intervals for the coefficients
estimated by logistic regression. The function uses resampling with replacement
to estimate the variability of the coefficient estimates.
}
\details{
This function takes the significance level, alpha, the number of
boostraps, beta, predictors, and observed variables (0 or 1) and computes a
more accurate confidence interval for the small data set, \code{beta_estimates}.
It creates a new larger sample, \code{boot_mean}, using values from the original
data set, \code{beta_estimates}. This larger sample is made creating \code{n}
sized samples, \code{beta_estimates_star}, and using the mean of each \code{n}
sized sample. The \code{n} sized samples are made by selecting numbers randomly
from the original data set with replacement. The function then uses the quantile
function to create the new confidence interval using the alpha provided by the
user.
}
\examples{
data(mtcars)
X <- mtcars$mpg
y <- mtcars$am
Bootstrap_function(alpha = 0.05, B = 20, X = X, y = y)
}
\author{
Shannon Clark
}
