% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group2.R
\name{cost_function}
\alias{cost_function}
\title{Cost Function for Logistic Regression}
\usage{
cost_function(beta, X, y)
}
\arguments{
\item{beta}{A \code{vector} of coefficients for the logistic regression model.}

\item{X}{A \code{matrix} containing the predictor variables. Each row corresponds
to an observation, and each column corresponds to a predictor variable.}

\item{y}{A \code{vector} representing the observed binary outcomes (0 or 1)
for each observation.}
}
\value{
A \code{numeric} value the total cost or loss calculated using the
cross-entropy loss formula. The return value indicates how well the modeled
coefficients fit the observed data. A better fit is represented by a lower value.
}
\description{
Computes the cost for logistic regression given coefficients,
predictor variables, and observed outcomes using the logistic function and
the cross-entropy loss formula.
}
\details{
The function calculates the predicted probabilities using the logistic function
and then computes the cost using the formula:
\code{-sum(y * log(p + 1e-9) + (1 - y) * log(1 - p + 1e-9))}.
The small constant (1e-9) is added to probabilities in the log function to avoid log(0).
}
\examples{
beta <- c(0.5, -0.25)
X <- matrix(c(1, 2, 3, 2, 1, 0), nrow = 3, byrow = TRUE)
y <- c(1, 0, 1)
cost_model <- cost_function(beta, X, y)
}
\author{
Cameron Tice
}
