% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group2.R
\name{cutoff_value_function}
\alias{cutoff_value_function}
\title{Determine Optimal Cutoff Value for Logistic Regression Predictions}
\usage{
cutoff_value_function(X, y, metric)
}
\arguments{
\item{X}{A \code{vector} or \code{matrix} representing the predictor variables.}

\item{y}{A \code{vector} representing the binary response variable.}

\item{metric}{A \code{string} specifying the performance metric to be evaluated.
Possible metrics include "accuracy", "sensitivity", "specificity",
"false discovery rate", "diagnostic odds ratio", and "prevalence".}
}
\value{
Generates a plot showing the performance of the specified metric
at different cutoff values ranging from 0.1 to 0.9. The plot helps in determining
the optimal cutoff value for making predictions.
}
\description{
Evaluates logistic regression predictions at different cutoff values
for a specified performance metric. The function provides a visual plot of the
performance metric across various cutoff thresholds.
}
\details{
Creates a \code{vector} for the metric chosen that has all of the values
for that metric for each cut-off value from 0.1 to 0.9 with increments of 0.1.
It plots this \code{vector} against the cut-off values.#'
}
\examples{
data(mtcars)
cutoff_value_function(X = mtcars$wt, y = mtcars$am, metric = "sensitivity")
}
\author{
Shannon Clark
}
