% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group2.R
\name{gradient_function}
\alias{gradient_function}
\title{Gradient Function for Logistic Regression}
\usage{
gradient_function(beta, X, y)
}
\arguments{
\item{beta}{A \code{vector} of coefficients for the logistic regression model.}

\item{X}{A \code{matrix} containing the predictor variables. Each row corresponds
to an observation, and each column corresponds to a predictor variable.}

\item{y}{A \code{vector} representing the observed binary outcomes (0 or 1)
for each observation.}
}
\value{
A \code{matrix} that represents the gradient of the cost function with
respect to the coefficients.
}
\description{
Computes the gradient of the cost function for logistic regression.
This function calculates the difference between the predicted probabilities and
the actual outcomes, scaled by the predictor variables.
}
\details{
Calculates the gradient as the product of the transposed matrix of
predictor variables and the vector of residuals. The gradient indicates the
direction and rate of fastest increase of the cost function, and is used to
update coefficients in the optimization process.
}
\examples{
beta <- c(0.5, -0.25)
X <- matrix(c(1, 2, 3, 2, 1, 0), nrow = 3, byrow = TRUE)
y <- c(1, 0, 1)
gradient_model <- gradient_function(beta, X, y)
}
\author{
Cameron Tice
}
