% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group2.R
\name{logistic_function}
\alias{logistic_function}
\title{Logistic Function}
\usage{
logistic_function(beta, X)
}
\arguments{
\item{beta}{A \code{vector} of coefficients that represent the relationship
between the predictor variable and the log-odds of the outcome.}

\item{X}{A \code{matrix} containing the predictor variables. Each row corresponds
to an observation, and each column corresponds to a predictor variable.}
}
\value{
A \code{vector} containing the estimated probability of the dependent
variable being a \code{1} or a \code{0}.
}
\description{
Computes logistic function for given predictors X and coefficients beta.
}
\examples{
X <- matrix(c(1, 2, 3, 4, 1, 0), nrow = 3, byrow = TRUE)
beta <- c(0.5, -0.2)
predicted_probabilities <- logistic_regression(beta, X)
}
\author{
Cameron Tice
}
