% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group2.R
\name{plot_logistic_curve}
\alias{plot_logistic_curve}
\title{Plot Logistic Regression Curve}
\usage{
plot_logistic_curve(X, y)
}
\arguments{
\item{X}{A \code{vector} or a single-column \code{matrix} representing the
predictor variables.}

\item{y}{A \code{vector} representing the binary response variable.}
}
\value{
A plot with the original data points and the fitted logistic regression
curve. Points are plotted in red and the logistic curve is plotted in blue.
}
\description{
Plots the logistic regression curve along with the original binary
response data points. The function fits a logistic regression model and then
uses the estimated coefficients to predict probabilities, which are plotted
as a curve.
}
\examples{
data(mtcars)
plot_logistic_curve(mtcars$wt, mtcars$am)
}
\author{
Cameron Tice
}
